Additional commands on MIPS

+++ Mips-net/Module/ITwaveModule.cs
@@ -7,14 +7,14 @@ namespace Mips.Module
 {
     public interface ITwaveModule
     {
-        Task<int> GetTWaveFrequency(string channel);
+        Task<double> GetTWaveFrequency(string channel);
         Task<Unit> SetTWaveFrequency(string channel, int frequency);
-        Task<int> GetTWavePulseVoltage(string channel);
+        Task<double> GetTWavePulseVoltage(string channel);
         Task<Unit> SetTWavePulseVoltage(string channel, int voltage);
         Task<Unit> SetTWaveGuard1Voltage(string channel, int voltage);
-        Task<int> GetTWaveGuard1Voltage(string channel);
+        Task<double> GetTWaveGuard1Voltage(string channel);
         Task<Unit> SetTWaveGuard2Voltage(string channel, int voltage);
-        Task<int> GetTWaveGuard2Voltage(string channel);
+        Task<double> GetTWaveGuard2Voltage(string channel);
         Task<BitArray> GetTWaveSequence(string channel);
         Task<Unit> SetTWaveSequence(string channel, BitArray sequence);
         Task<TWaveDirection> GetTWaveDirection(string channel);
@@ -25,13 +25,13 @@ namespace Mips.Module
         Task<Unit> SetCompressorMode(CompressorMode mode);
         Task<int> GetCompressorOrder();
         Task<Unit> SetCompressorOrder(int order);
-        Task<int> GetCompressorTriggerDelay();
-        Task<Unit> SetCompressorTriggerDelay(int delayMilliseconds);
-        Task<int> GetCompressionTime();
+        Task<double> GetCompressorTriggerDelay();
+        Task<Unit> SetCompressorTriggerDelay(double delayMilliseconds);
+        Task<double> GetCompressionTime();
         Task<Unit> SetCompressionTime(int timeMilliseconds);
-        Task<int> GetNormalTime();
+        Task<double> GetNormalTime();
         Task<Unit> SetNormalTime(int timeMilliseconds);
-        Task<int> GetNonCompressTime();
+        Task<double> GetNonCompressTime();
         Task<Unit> SetNonCompressTime(int timemilliSeconds);
         Task<Unit> ForceMultipassTrigger();
         Task<SwitchState> GetSwitchState();
